refactor: auth service
